fixed some more msvc compiler warnings

+++ code/api/test/user/data/grid.cc
@@ -714,7 +714,7 @@ namespace data {
 		// fill in some data
 		auto dataSrc = src.mask();
 		full.scan([&](const GridPoint<1>& p){
-			src[p] = p[0];
+			src[p] = (int)p[0];
 		});
 
 
@@ -776,7 +776,7 @@ namespace data {
 		// fill in some data
 		auto dataSrc = src.mask();
 		full.scan([&](const GridPoint<2>& p){
-			src[p] = p[0] * p[1];
+			src[p] = (int)(p[0] * p[1]);
 		});
